# target/generated-sources/openapi/src/main/java/com/portfolio/users/generated/api/UsersApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.5.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.portfolio.users.generated.api;

import com.portfolio.users.generated.model.CreateUserRequest;
import java.util.UUID;
import com.portfolio.users.generated.model.UpdateUserRequest;
import com.portfolio.users.generated.model.User;
import com.portfolio.users.generated.model.UserExists200Response;
import com.portfolio.users.generated.model.UserPage;
import com.portfolio.users.generated.model.UserStatus;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-06T08:25:35.138267500+01:00[Europe/Madrid]", comments = "Generator version: 7.5.0")
@Validated
@Tag(name = "Users", description = "People interacting with the portfolio platform.")
public interface UsersApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /users : Create user
     *
     * @param createUserRequest  (required)
     * @return Created (status code 201)
     *         or Email already registered (status code 409)
     */
    @Operation(
        operationId = "createUser",
        summary = "Create user",
        tags = { "Users" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = User.class))
            }),
            @ApiResponse(responseCode = "409", description = "Email already registered")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/users",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<User> createUser(
        @Parameter(name = "CreateUserRequest", description = "", required = true) @Valid @RequestBody CreateUserRequest createUserRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"skills\" : [ \"skills\", \"skills\" ], \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"phoneNumber\" : \"phoneNumber\", \"fullName\" : \"fullName\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"headline\" : \"headline\", \"email\" : \"email\", \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /users/{id} : Soft delete user
     *
     * @param id  (required)
     * @return Deleted (status code 204)
     */
    @Operation(
        operationId = "deleteUser",
        summary = "Soft delete user",
        tags = { "Users" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Deleted")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/users/{id}"
    )
    
    default ResponseEntity<Void> deleteUser(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /users/{id} : Get user by id
     *
     * @param id  (required)
     * @return Found (status code 200)
     *         or Missing (status code 404)
     */
    @Operation(
        operationId = "getUser",
        summary = "Get user by id",
        tags = { "Users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = User.class))
            }),
            @ApiResponse(responseCode = "404", description = "Missing")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/users/{id}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<User> getUser(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"skills\" : [ \"skills\", \"skills\" ], \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"phoneNumber\" : \"phoneNumber\", \"fullName\" : \"fullName\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"headline\" : \"headline\", \"email\" : \"email\", \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /users/email/{email} : Fetch user by email
     *
     * @param email  (required)
     * @return Found (status code 200)
     *         or Missing (status code 404)
     */
    @Operation(
        operationId = "getUserByEmail",
        summary = "Fetch user by email",
        tags = { "Users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = User.class))
            }),
            @ApiResponse(responseCode = "404", description = "Missing")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/users/email/{email}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<User> getUserByEmail(
        @jakarta.validation.constraints.Email @Parameter(name = "email", description = "", required = true, in = ParameterIn.PATH) @PathVariable("email") String email
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"skills\" : [ \"skills\", \"skills\" ], \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"phoneNumber\" : \"phoneNumber\", \"fullName\" : \"fullName\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"headline\" : \"headline\", \"email\" : \"email\", \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /users : List users
     *
     * @param page  (optional)
     * @param size  (optional)
     * @param status  (optional)
     * @return Page of users (status code 200)
     */
    @Operation(
        operationId = "listUsers",
        summary = "List users",
        tags = { "Users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Page of users", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserPage.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/users",
        produces = { "application/json" }
    )
    
    default ResponseEntity<UserPage> listUsers(
        @Min(0) @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false) Integer page,
        @Min(1) @Max(100) @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false) Integer size,
        @Parameter(name = "status", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "status", required = false) UserStatus status
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"size\" : 6, \"totalPages\" : 5, \"page\" : 0, \"content\" : [ { \"skills\" : [ \"skills\", \"skills\" ], \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"phoneNumber\" : \"phoneNumber\", \"fullName\" : \"fullName\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"headline\" : \"headline\", \"email\" : \"email\", \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }, { \"skills\" : [ \"skills\", \"skills\" ], \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"phoneNumber\" : \"phoneNumber\", \"fullName\" : \"fullName\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"headline\" : \"headline\", \"email\" : \"email\", \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" } ], \"totalElements\" : 1 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /users/{id} : Update user
     *
     * @param id  (required)
     * @param updateUserRequest  (required)
     * @return Updated (status code 200)
     *         or Missing (status code 404)
     */
    @Operation(
        operationId = "updateUser",
        summary = "Update user",
        tags = { "Users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Updated", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = User.class))
            }),
            @ApiResponse(responseCode = "404", description = "Missing")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/users/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<User> updateUser(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id,
        @Parameter(name = "UpdateUserRequest", description = "", required = true) @Valid @RequestBody UpdateUserRequest updateUserRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"skills\" : [ \"skills\", \"skills\" ], \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"phoneNumber\" : \"phoneNumber\", \"fullName\" : \"fullName\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"headline\" : \"headline\", \"email\" : \"email\", \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /users/{id}/exists : Check if user exists
     *
     * @param id  (required)
     * @return Boolean flag (status code 200)
     */
    @Operation(
        operationId = "userExists",
        summary = "Check if user exists",
        tags = { "Users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Boolean flag", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserExists200Response.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/users/{id}/exists",
        produces = { "application/json" }
    )
    
    default ResponseEntity<UserExists200Response> userExists(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"exists\" : true }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}